Move static ExceptionMessages class to UI package
