+++ src/main/java/chessmaster/exceptions/InvalidMoveException.java
@@ -1,5 +1,7 @@
 package chessmaster.exceptions;
 
+import chessmaster.ui.ExceptionMessages;
+
 public class InvalidMoveException extends ChessMasterException {
     public InvalidMoveException() {
         super(ExceptionMessages.MESSAGE_INVALID_MOVE_EXCEPTION);
