+++ src/main/java/chessmaster/exceptions/LoadBoardException.java
@@ -1,5 +1,7 @@
 package chessmaster.exceptions;
 
+import chessmaster.ui.ExceptionMessages;
+
 public class LoadBoardException extends ChessMasterException {
 
     public LoadBoardException() {
