+++ src/main/java/chessmaster/exceptions/MoveOpponentPieceException.java
@@ -1,5 +1,7 @@
 package chessmaster.exceptions;
 
+import chessmaster.ui.ExceptionMessages;
+
 public class MoveOpponentPieceException extends ChessMasterException {
     public MoveOpponentPieceException() {
         super(ExceptionMessages.MESSAGE_MOVE_OPPONENT_EXCEPTION);
