+++ src/main/java/chessmaster/exceptions/NullPieceException.java
@@ -1,5 +1,7 @@
 package chessmaster.exceptions;
 
+import chessmaster.ui.ExceptionMessages;
+
 public class NullPieceException extends ChessMasterException {
     public NullPieceException() {
         super(ExceptionMessages.MESSAGE_NULL_PIECE_EXCEPTION);
