+++ src/main/java/chessmaster/exceptions/ParseColorException.java
@@ -1,5 +1,7 @@
 package chessmaster.exceptions;
 
+import chessmaster.ui.ExceptionMessages;
+
 public class ParseColorException extends ChessMasterException {
     public ParseColorException() {
         super(ExceptionMessages.MESSAGE_PARSE_COLOR_EXCEPTION);
