+++ src/main/java/chessmaster/exceptions/SaveBoardException.java
@@ -1,5 +1,7 @@
 package chessmaster.exceptions;
 
+import chessmaster.ui.ExceptionMessages;
+
 public class SaveBoardException extends ChessMasterException {
 
     public SaveBoardException() {
